Create Products  Category Html .Male
This category has 4 columns and 1 row, which has 4 pictures. You need to change the pictures and the product's description.
There is a modal with 5 pictures. The first large picture you must insert the picture same as the one placed on each column and the 4 small pictures you must put the different views of the picture displayed. Then you must change the price, product name and product details.